No longer hardcodes number of iterations
Doing so means that you will likely miss some of the earlier stories (even if you edit the script right before running it.)  You can test this easily (as I did) by adding a tag for author_[some username with small number of stories] and setting hitsPerPage smaller than that number, and verifying that your output includes the oldest story.  It might also make sense to include a commandline option that does that switches into that mode for you.